dnn_from_scratch_experiments::mnist_experiment::dataset_setup
Function to_array3

fn to_array3<T, D>(array: ArrayBase<T, D>) -> ArrayBase<T, Ix3>
where
    T: Data,
    D: Dimension,

Converts an ndarray of any dimension (1D, 2D, or 3D) into a 3D array.

Arguments

array: An ArrayBase of any dimension to be converted to 3D.

Returns

A 3-dimensional array (ArrayBase<T, Ix3>).

Panics

Panics if the input array has a dimensionality other than 1D, 2D, or 3D.